Use Echo binding

+++ components/movies_templ.go
@@ -180,7 +180,7 @@ func ratingClass(rating float64) string {
 	}
 }
 
-func More(more bool, limit, offset int) templ.Component {
+func More(more bool, limit, offset uint64) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, templ_7745c5c3_W io.Writer) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_Buffer, templ_7745c5c3_IsBuffer := templ_7745c5c3_W.(*bytes.Buffer)
 		if !templ_7745c5c3_IsBuffer {
@@ -220,7 +220,7 @@ func More(more bool, limit, offset int) templ.Component {
 	})
 }
 
-func moreButton(limit, offset int) templ.Component {
+func moreButton(limit, offset uint64) templ.Component {
 	return templ.ComponentFunc(func(ctx context.Context, templ_7745c5c3_W io.Writer) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_Buffer, templ_7745c5c3_IsBuffer := templ_7745c5c3_W.(*bytes.Buffer)
 		if !templ_7745c5c3_IsBuffer {
